requirements.txt
streamlit>=1.55.0
google-genai>=1.0.0
python-dotenv>=1.2.2
requests>=2.33.0
